Move traits instantiation to a separate header (#93)
* Minor improvements.
* Also fix some wrong variable names in the Windows part of the handlers.
